+++ modules/hosts/caelestia/home.nix
@@ -33,6 +33,8 @@
                             self.homeModules.dolphin
                             self.homeModules.protonup-qt
                             self.homeModules.handbrake
+                            self.homeModules.fish
+                            self.homeModules.starship
                             self.homeModules.caelestiaPlatform
                         ];