+++ modules/experimental/_hosts/caelestia/symlinks.nix
@@ -1,0 +1,19 @@
+{ experimental, config, lib, pkgs, ... }:
+
+{
+    config = lib.mkIf experimental {
+        home.file = {
+            "./wallpapers" = {
+                source = "${pkgs.wallpapers}/share/wallpapers";
+                recursive = true;
+                force = true;
+            };
+
+            "./pfps" = {
+                source = "${pkgs.pfps}/share/pfps";
+                recursive = true;
+                force = true;
+            };
+        };
+    };
+}