+++ modules/experimental/_platforms/caelestia/caelestia-shell.nix
@@ -1,0 +1,26 @@
+{ experimental, config, inputs, ... }:
+
+{
+    imports = [
+        inputs.caelestia-shell.homeManagerModule.default
+    ];
+
+    programs.calestia = {
+        enable = true;
+        settings = {
+            bar.statusIcons = [
+                { id = "lockStatus"; enabled = true; }
+                { id = "network"; enabled = true; }
+            ];
+
+            paths.wallpaperDir = "~/wallpapers";
+        };
+
+        cli = {
+            enable = true;
+            settings.theme = {
+                enableGtk = false;
+            };
+        };
+    };
+}