+++ modules/experimental/default.nix
@@ -2,11 +2,15 @@
 
 let 
     experimentalNixos = {
+        calestiaHost = ./_hosts/caelestia/default.nix;
         open-webGui = ./_features/open-webgui.nix;
         ollama = ./_features/ollama.nix;
     };
 
-    experimentalHome = {};
+    experimentalHome = {
+        caelestiaPlatform = ./_platforms/caelestia/default.nix;
+        symlinks.nix = ./_hosts/caelestia/symlinks.nix;
+    };
 in 
 {
     flake.experimentalModules = {