+++ flake.nix
@@ -78,6 +78,13 @@
             flake = false;
         };
 
+        caelestia-shell = {
+            url = "github:caelestia-dots/shell";
+            inputs = {
+                nixpkgs.follows = "nixpkgs";
+            };
+        };
+
         niri.url = "github:sodiboo/niri-flake";
         flake-parts.url = "github:hercules-ci/flake-parts";
         nix-wrapper-modules.url = "github:BirdeeHub/nix-wrapper-modules";