+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added caelestia host.
 
 #### Changed
+- Removed `loadModules` list from the ollama `experimentalModule`.
 - Updated `architecture.md` to document both the caelestia host and `caelestiaPlatform`.
 - Disabled experimental features within `illogical-impulseConfiguration` and `illogical-impulseHome`.
 - Renamed `tempmonitors.nix` to `tempconfig.nix`.