+++ CHANGELOG.md
@@ -2,6 +2,42 @@
 
 ## [Unreleased]
 
+
+## [v0.5.0]
+
+### 2026-09-18
+
+#### Added
+- Added `experimentalFeatures.md` to explain `experimentalModules` further, alongside the enabiling process.
+- Added `eamodio.gitlens` extension to `vscodium` homeModule.
+
+#### Changed
+- Updated `architecture.md` to add a link to `experimentalFeatures.md`.
+- Updated `architecture.md` to note that caelestia is experimental.
+- [Breaking] Converted both the caelestia host and caelestia platform to `experimentalModules`, therefore the platform is disabled by default and requires enabiling.
+
+### 2026-09-13
+
+#### Added
+- Imported `fish` and `starship` into `caelestiaHome`.
+- Added `starship` homeModule.
+- Added `fish` homeModule.
+- Imported `caelestiaPlatform` into `caelestiaHome`.
+- Added `caelestiaPlatform`.
+- Added `caelestia-shell` input (`flake.nix`).
+- Added caelestia host.
+
+#### Changed
+- Removed `loadModules` list from the ollama `experimentalModule`.
+- Updated `architecture.md` to document both the caelestia host and `caelestiaPlatform`.
+- Disabled experimental features within `illogical-impulseConfiguration` and `illogical-impulseHome`.
+- Renamed `tempmonitors.nix` to `tempconfig.nix`.
+
+#### Fixed
+- Corrected temporary hyprland configuration file name within `tempconfig.nix`.
+- Corrected module name within `end4-pCPlatform`.
+
+
 ## [v0.4.1]
 
 ### - 2026-09-11